shuffled functions around for bytebudget
