+++ decompile/General/AltMods/Mods4.c
@@ -1,5 +1,6 @@
 #ifdef USE_ONLINE
 #include "OnlineCTR/names3d.c"
+#include "OnlineCTR/misc_states2.c"
 
 typedef void (*VehicleFuncPtr)(struct Thread* thread, struct Driver* driver);
 
@@ -56,9 +57,7 @@ void RunVehicleSet13(struct Thread* dThread, struct Driver* dOnline)
 		#endif
 	}
 }
-#endif
 
-#ifdef USE_ONLINE
 Color HsvToRgb(int h, int s, int v)
 {
     Color rgb;
