+++ decompile/General/AltMods/Mods6.c
@@ -1,0 +1,3 @@
+#if defined(USE_ONLINE)
+#include "OnlineCTR/timestamp.c"
+#endif
