+++ decompile/General/AltMods/Mods8.c
@@ -1,3 +1,4 @@
-#ifdef USE_ONLINE
+#if defined(USE_ONLINE)
 #include "OnlineCTR/lap.c"
+#include "OnlineCTR/misc_states3.c"
 #endif
